OK-746: Refaktoroitu erilleen yleiskäyttöinen FileUploadButton-komponentti
